fix: XXXXX_mark_reactionの潜在的なバグを排除
以前のコードではXXXXX_mark_reactionが付与されていない場合のことを考慮しているように見えますが、
本当にXXXXX_mark_reactionが付与されていない場合、XXXXX_mark_reaction.countはNameErrorになります。
has_prospect() == True かつ XXXXX_mark_reaction == None は本来ありえないため、このバグは今まで発現しませんでした。

+++ cogs/approval.py
@@ -48,7 +48,7 @@ class Approval:
                 if reaction.emoji == "✅":
                     check_mark_reaction = reaction
 
-            if check_mark_reaction.count > 3:
+            if check_mark_reaction and check_mark_reaction.count > 3:
                 role = discord.utils.find(lambda m: m.name == 'player', user.guild.roles)
                 await message.author.add_roles(role)
                 role = discord.utils.find(lambda m: m.name == 'prospect', user.guild.roles)
@@ -67,7 +67,7 @@ class Approval:
                 if reaction.emoji == "❎":
                     x_mark_reaction = reaction
 
-            if x_mark_reaction.count > 3:
+            if x_mark_reaction and x_mark_reaction.count > 3:
                 await message.channel.send(f"{message.author.mention} 申請が否認されました")
                 await message.delete()
             else:
